make zapd addresses globals int64_t so they can store uint32_t addresses and -1

+++ tools/ZAPD/ZAPD/Globals.h
@@ -39,9 +39,9 @@ public:
 	fs::path baseRomPath, inputPath, outputPath, sourceOutputPath, cfgPath;
 	TextureType texType;
 	CsFloatType floatType = CsFloatType::FloatOnly;
-	int baseAddress = -1;
-	int startOffset = -1;
-	int endOffset = -1;
+	int64_t baseAddress = -1;
+	int64_t startOffset = -1;
+	int64_t endOffset = -1;
 	ZGame game;
 	GameConfig cfg;
 	bool verboseUnaccounted = false;
